backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.3
fastapi==0.124.0
h11==0.16.0
idna==3.11
pdfminer.six==20251107
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
python-multipart==0.0.20
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
